add apache commons CollectionUtils and MapUtils
